+++ src/JSONs.js
@@ -1,12 +1,14 @@
 /**
  * json-strictify
  *
- * @version 0.2.1
+ * @version 0.3.0
  * @author Raphael Pigulla <pigulla@four66.com>
  * @license MIT
  */
 
 var util = require('util');
+
+var fkt = require('fkt');
 
 var CircularReferenceError = require('./CircularReferenceError'),
     InvalidValueError = require('./InvalidValueError');
@@ -19,7 +21,7 @@ var JSONs = {
 
     /**
      * Recursively check if the given object can be serialized to JSON safely.
-     * 
+     *
      * @param {Object} object
      * @param {Array.<(string|number)>} references
      * @param {Array.<(Object|Array)>} ancestors
@@ -36,19 +38,19 @@ var JSONs = {
             actual = object.toJSON();
             return this.check(actual, references, ancestors);
         }
-    
+
         for (var key in object) { // jshint ignore:line
             actual = this.replacer ? this.replacer(key, object[key]) : object[key];
-    
+
             if (!(this.replacer && actual === undefined)) {
                 this.check(actual, references.concat(key), ancestors.concat(object));
             }
         }
     },
-    
+
     /**
      * Recursively check if the given array can be serialized to JSON safely.
-     * 
+     *
      * @param {Array} array
      * @param {Array.<(string|number)>} references
      * @param {Array.<(Object|Array)>} ancestors
@@ -64,7 +66,7 @@ var JSONs = {
             this.check(actual, references.concat(index), ancestors.concat([array]));
         }, this);
     },
-    
+
     /**
      * Check if the given value is of a known, not-serializable type and provide a more specific, helpful error message.
      *
@@ -91,10 +93,10 @@ var JSONs = {
             throw new InvalidValueError(value + ' is not a valid JSON type', value, references);
         }
     },
-    
+
     /**
      * Recursively check if the given value can be serialized to JSON safely.
-     * 
+     *
      * @param {*} value
      * @param {Array.<(string|number)>} references
      * @param {Array.<(Object|Array)>} ancestors
@@ -105,12 +107,12 @@ var JSONs = {
     check: function (value, references, ancestors) {
         // Check for the most common non-serializable types.
         this.checkCommonTypes(value, references);
-    
+
         // Primitive types are always okay (we've already checked for non-finite numbers).
         if (value === null || typeof value === 'string' || typeof value === 'boolean' || typeof value === 'number') {
             return;
         }
-    
+
         if (Array.isArray(value)) {
             // If an array, check its elements.
             return this.checkArray(value, references, ancestors);
@@ -123,17 +125,17 @@ var JSONs = {
             throw new InvalidValueError('Invalid type', value);
         }
     },
-    
+
     /**
      * Normalizes the user-specified replacer function.
-     * 
+     *
      * In short, JSON.stringify's "replacer" parameter can either be a function or an array containing the names of the
      * properties to be included. This method normalizes the latter case to the former so we can always treat the
      * "replacer" option as a function internally.
-     * 
+     *
      * For more information about the replacer function take a look at the documentation on
      * [MDN](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Using_native_JSON#The_replacer_parameter).
-     * 
+     *
      * @param {(function(string,*)|Array.<(string|number)>)=} replacer
      * @return {?function(string,*):*}
      */
@@ -143,13 +145,13 @@ var JSONs = {
                 return (key !== '' && replacer.indexOf(key) === -1) ? undefined : value;
             };
         }
-    
+
         return typeof replacer === 'function' ? replacer : null;
     },
 
     /**
      * Check if the passed value is a circular reference, i.e. whether it is one of its own ancestors.
-     * 
+     *
      * @param {(Object|Array)} value
      * @param {Array.<(string|number)>} references
      * @param {Array.<(Object|Array)>} ancestors
@@ -160,10 +162,10 @@ var JSONs = {
             throw new CircularReferenceError(references);
         }
     },
-    
+
     /**
      * The drop-in replacement function for JSON.stringify.
-     * 
+     *
      * @param {*} value
      * @param {(function(string,*):*|Array.<(string|number)>)=} replacer
      * @param {number=} space
@@ -177,18 +179,33 @@ var JSONs = {
         var initialData = this.replacer ? this.replacer('', value) : value;
 
         this.check(initialData, [], []);
-        
+
         // Fall back to the native JSON.stringify that we now now is safe to use.
         return JSON.stringify(value, replacer, space);
     }
 };
 
-var wrapper = {
+var nativeImpl,
+    strictImpl;
+
+nativeImpl = {
     parse: JSON.parse,
-    stringify: JSONs.stringify.bind(JSONs),
-    enable: function (enabled) {
-        return enabled ? wrapper : JSON;
+    parseAsync: fkt.callbackify(JSON.parse, JSON),
+    stringify: JSON.stringify,
+    stringifyAsync: fkt.callbackify(JSON.stringify, JSON),
+    enabled: function (enabled) {
+        return enabled ? strictImpl : nativeImpl;
     }
 };
 
-module.exports = wrapper;
+strictImpl = {
+    parse: JSON.parse,
+    parseAsync: fkt.callbackify(JSON.parse, JSON),
+    stringify: JSONs.stringify.bind(JSONs),
+    stringifyAsync: fkt.callbackify(JSONs.stringify, JSONs),
+    enabled: function (enabled) {
+        return enabled ? strictImpl : nativeImpl;
+    }
+};
+
+module.exports = strictImpl;