+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0 (2014-10-14)
+
+ - Breaking Change: `enable` is now called `enabled`
+ - added `parseAsync` and `stringifyAsync`
+
 ## 0.2.1 (2014-10-10)
 
  - maintenance release